project/scripts/requirements.txt
pydicom
Pillow
requests
numpy